requirements.txt
numpy==1.26.0
opencv-python==4.8.1
Keras==2.14.0
Pillow==10.1.0
tensorflow==2.14.0
